# doc/api.yaml
openapi: 3.0.3
info:
  title: WaSAText API
  description: This is the API for the first WaSAText Homework.
  version: 1.0.0
...
components:
  responses:
    BadRequest:
      description: Bad Request - Invalid input data.
      content:
        application/json:
          schema:
            type: object
            description: Represents an error response.
            properties:
              errorMessage:
                type: string
                example: "Bad Request"
                description: The server could not process the request due to invalid syntax or missing required parameters.
    Unauthorized:
      description: Unauthorized - The user is not authenticated.
      content:
        application/json:
          schema:
            type: object
            description: Represents an error response.
            properties:
              errorMessage:
                type: string
                example: "Unauthorized"
                description: Authentication is required and has failed or has not been provided.
    Forbidden:
      description: Forbidden - The action is prohibited.
      content:
        application/json:
          schema:
            type: object
            description: Represents an error response.
            properties:
              errorMessage:
                type: string
                example: "Forbidden"
                description: The client does not have permission to access the requested resource.
    NotFound:
      description: Not Found - The resource could not be found.
      content:
        application/json:
          schema:
            type: object
            description: Represents an error response.
            properties:
              errorMessage:
                type: string
                example: "Not Found"
                description: The server cannot find the requested resource. The URL may be incorrect or the resource does not exist.
    InternalServerError:
      description: Internal Server Error - An unexpected server error occurred.
      content:
        application/json:
          schema:
            type: object
            description: Represents an error response.
            properties:
              errorMessage:
                type: string
                example: "Internal Server Error"
                description: The server encountered an unexpected condition that prevented it from fulfilling the request.
  schemas:
    username:
      type: string
      minLength: 3
      maxLength: 16
      pattern: '^.*?$'
      example: Maria
      description: The username which is between 3 and 16 characters long.
...
    userId:
      type: integer
      example: 1
      description: The unique identifier of the user.
    chatId:
      type: integer
      example: 2
      description: The unique identifier of the conversation.
...
  securitySchemes:
    securityKey:
      type: apiKey
      in: header
      name: authorizationKey
      description: API key that will be used for authorization.
paths:
...
  /users/{id}/username:
    parameters:
      - name: id
        in: path
        description: The ID of the user to be updated.
        required: true
        schema: { $ref: '#/components/schemas/userId' }
        
    put:
      tags: ['users']
      summary: Set the user's display name
      description: Sets a new username for the specified user.
      operationId: setMyUserName
      security:
        - securityKey: []
      requestBody:
        description: The new username for the specified user.
        content:
          application/json:
            schema:
              type: object
              description: Updated username.
              properties:
                username: { $ref: '#/components/schemas/username' }
              required:
                - username
        required: true
      responses:
        '204': { description:  Successfully updated the username. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' } 
          
    get:
      tags: ['users']
      summary: Retrieve the user's name
      description: Retrieve the username via user ID.
      operationId: getUsername
      security:
        - securityKey: []
      responses:
        '200':
          description: Successfully retrieved the username.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/username' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' } 
...
  /newchat:
    put:
      tags: ["groups"]
      summary: Begin a conversation with other users
      description: Allows the current user to start a conversation with 1 or more users.
      operationId: newChat
      security:
        - securityKey: []
      requestBody:
        description: Details about the new conversation.
        content:
          application/json:
            schema:
              description: The IDs of the chat members
              type: object
              properties:
                members:
                  type: array
                  items: { $ref: '#/components/schemas/userId' }
                  minItems: 1
                  maxItems: 2000
                  example: [1, 12, 123]
                  description: User IDs to be added to the new chat.
              required:
                - members
        required: true
      responses:  
        '200':
          description: Successfully created a new conversation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/chatId' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }  
...